[PINOT-2912]: Aggregation GroupBy Rewrite: Cleanup: Merged SV/MV group by executors, eliminated map based GroupKeyGenerator and ResultHolder.
RB=710653
G=pinot-dev-reviewers
R=kgopalak,jfim,atumbde,ssubrama
A=kgopalak

+++ pinot-core/src/main/java/com/linkedin/pinot/core/operator/groupby/DoubleArrayBasedResultHolder.java
@@ -64,8 +64,8 @@ public class DoubleArrayBasedResultHolder implements ResultHolder {
    * @return
    */
   @Override
-  public double getDoubleResult(long groupKey) {
-    return _resultArray.getDoubleResult((int) groupKey);
+  public double getDoubleResult(int groupKey) {
+    return _resultArray.getDoubleResult(groupKey);
   }
 
   @Override
